#172: ProductDateChoiceTableViewCell 파일 분리
